wf(phase 4): Fold the six overlapping pages into their counterparts
