+++ projects/gnrcore/packages/test/tests/docstring_debt.txt
@@ -5,14 +5,11 @@ test/webpages/components/Grid/evaluation_grid.py
 test/webpages/components/TableHandler/borderTableHandler.py
 test/webpages/components/document_frame.py
 test/webpages/components/full_calendar.py
-test/webpages/components/includedview.py
-test/webpages/components/multibutton.py
 test/webpages/components/paged_editor.py
 test/webpages/components/palette_importer.py
 test/webpages/components/templateChunk.py
 test/webpages/components/text_area.py
 test/webpages/components/tpleditor.py
-test/webpages/dojo/menu.py
 test/webpages/gnrwdg/flexbox.py
 test/webpages/gnrwdg/formbuilder.py
 test/webpages/gnrwdg/test_grouplet_grid/04_setlayout_dnd.py
@@ -52,16 +49,11 @@ test15/webpages/chart/chartjs.py
 test15/webpages/chart/dygraph.py
 test15/webpages/dd/dd_grid.py
 test15/webpages/events/keyboard_logger.py
-test15/webpages/gnrwdg/baggrid_scroll.py
 test15/webpages/gnrwdg/formHandler.py
 test15/webpages/gnrwdg/formclientstore.py
 test15/webpages/gnrwdg/formhandler_baggrid.py
 test15/webpages/gnrwdg/formhandler_selection.py
-test15/webpages/gnrwdg/framepane.py
 test15/webpages/gnrwdg/gridscale.py
-test15/webpages/gnrwdg/menuselect.py
-test15/webpages/gnrwdg/palette.py
-test15/webpages/gnrwdg/slotbar.py
 test15/webpages/gnrwdg/thpalette.py
 test15/webpages/html/test_hidden.py
 test15/webpages/html/webRTC.py
